backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
motor==3.7.0
pydantic==2.10.6
pydantic-settings==2.7.1
firebase-admin==6.7.0
httpx==0.28.1
numpy==2.2.2
pillow==11.1.0
scikit-learn==1.8.0
pandas==3.0.1
python-dotenv==1.0.1
torch
torchvision
